Fixes #14
Addition input files added.
Issue appears to have been caused by one pathological instance of a compound that
resulted in a baseline mortality = 1.0, which resulted in an infinite value to be
returned to the gsl library, and things got ugly. Fixed through two methods:
random perturbation of initial parameter guess, and also through resetting the
initial guess of background mortality to 0.99 if |base_mort - 1.0| < 1e-6.