Resolves #15
Added accept/cancel change buttons to data previewer. Datapreviewer now takes
parent as argument instead of parts of parent, and deep copies of the UIDH and
the stats_obj are made. All changes are kept to these deep copies, unless
changes are accepted, at which point the parent's references are updated and
window is destroyed.